+++ lib/spicerack/configurable/config_builder.rb
@@ -3,27 +3,39 @@
 module Spicerack
   module Configurable
     class ConfigBuilder
+      delegate :config_eval, to: :reader
+
       def reader
         @reader ||= Reader.new(configuration)
       end
 
       def configure
-        yield configuration
+        mutex.synchronize do
+          yield configuration
+        end
       end
 
       # NOTE: options must be set up before {#configure} is called
-      def option(*args)
-        config_class.__send__(:option, *args)
+      def option(*args, &block)
+        config_class.__send__(:option, *args, &block)
+      end
+
+      def nested(*args, &block)
+        config_class.__send__(:nested, *args, &block)
       end
 
       private
 
       def configuration
-        @configuration ||= config_class.new
+        config_class.instance
       end
 
       def config_class
-        @config_class ||= Class.new(InputObject)
+        @config_class ||= Class.new(ConfigObject)
+      end
+
+      def mutex
+        @mutex = Mutex.new
       end
     end
   end