+++ lib/spicerack/configurable/reader.rb
@@ -7,18 +7,31 @@ module Spicerack
         @config = config
       end
 
+      def config_eval(*path)
+        Evaluator.new(path, self)
+      end
+
       private
 
       attr_reader :config
 
       def method_missing(method_name, *)
-        return config.public_send(method_name) if config._options.map(&:to_sym).include?(method_name.to_sym)
+        name = method_name.to_sym
+
+        return mutex.synchronize { config.public_send(name) } if config._options.map(&:to_sym).include?(name)
+        return config._nested_builders[name].reader if config._nested_builders.key?(name)
 
         super
       end
 
       def respond_to_missing?(method_name, *)
-        config._options.map(&:to_sym).include?(method_name.to_sym) || super
+        config._options.map(&:to_sym).include?(method_name.to_sym) ||
+          config._nested_builders.key?(name) ||
+          super
+      end
+
+      def mutex
+        @mutex ||= Mutex.new
       end
     end
   end